+++ includes/ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: slindgre <slindgre@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/01 20:57:11 by gloras-t          #+#    #+#             */
-/*   Updated: 2020/12/20 23:47:39 by slindgre         ###   ########.fr       */
+/*   Updated: 2020/12/21 02:00:16 by slindgre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,5 +63,10 @@ int		print_files(t_list *files, uint8_t options, int mode);
 */
 t_list	*read_dir(t_file *dir, uint8_t options);
 void	print_dirs(t_list *dirs, uint8_t options, int step);
+
+/*
+** utils.c
+*/
+void	get_file_mtime(t_file *file, char *time_str);
 
 #endif
